Use rug::Float instead of f64

+++ backend/libs/exchanges/executors/src/binance/front_test.rs
@@ -5,6 +5,7 @@ use ::async_trait::async_trait;
 use ::futures::stream::{BoxStream, StreamExt};
 use ::mongodb::bson::oid::ObjectId;
 use ::nats::jetstream::JetStream as NatsJS;
+use ::rug::Float;
 
 use ::entities::{
   BookTicker, ExecutionSummary, ExecutionType, Order, OrderInner, OrderOption,
@@ -24,12 +25,16 @@ pub struct Executor {
   orders: HashMap<ObjectId, Order>,
   positions: HashMap<ObjectId, OrderInner>,
   cur_trade: Option<BookTicker>,
-  maker_fee: f64,
-  taker_fee: f64,
+  maker_fee: Float,
+  taker_fee: Float,
 }
 
 impl Executor {
-  pub async fn new(broker: &NatsJS, maker_fee: f64, taker_fee: f64) -> Self {
+  pub async fn new(
+    broker: &NatsJS,
+    maker_fee: Float,
+    taker_fee: Float,
+  ) -> Self {
     return Self {
       observer: TradeObserver::new(None, broker).await,
       orders: HashMap::new(),
@@ -46,10 +51,10 @@ impl TestExecutorTrait for Executor {
     return self.cur_trade.clone();
   }
 
-  fn maker_fee(&self) -> f64 {
+  fn maker_fee(&self) -> Float {
     return self.maker_fee;
   }
-  fn taker_fee(&self) -> f64 {
+  fn taker_fee(&self) -> Float {
     return self.taker_fee;
   }
   fn get_orders(&self) -> HashMap<ObjectId, Order> {
@@ -88,8 +93,8 @@ impl ExecutorTrait for Executor {
     &mut self,
     _: ObjectId,
     _: String,
-    _: Option<f64>,
-    _: f64,
+    _: Option<Float>,
+    _: Float,
     _: Option<OrderOption>,
   ) -> ExecutionResult<ObjectId> {
     return Err(
